Store the pointer to TipWindow in GraphCtrl instead of using IDs
Using FindWindow(ID_TIPWIN) stopped working in wxWidgets 3.0, as
FindWindow() now skips top-level children of the window. And even when
it did work (or if it starts working again), it's still simpler to avoid
using it and just store a pointer to the TipWindow when creating it and
set it to null when destroying it.

This makes tooltips hiding work with wxWidgets 3.x too.