Correct tip window position when not on the main display
The tip window position was always made at least 0, but this was wrong
as a display positioned to the left of the main one can perfectly well
use negative coordinates.

The code checking that the tooltip fit on the display also always used
the main display size, which was again incorrect for multiple displays.

Fix both of these problems.