Fix build with Graphviz 2.38
Graphviz versions after 2.30 always use cgraph library, which doesn't
provide direct access to the object name.

Use agnameof() function instead to retrieve it.

Notice that while this allows the code to work with both old and new
versions of Graphviz with only a small modification, this is not the
best way to do it with current Graphviz versions as they allow using IDs
to directly identify the graph elements instead of encoding their
addresses in the names as we do currently. When support for pre-cgraph
Graphviz versions is dropped in the future, we should consider changing
this code to use the IDs instead for simplicity and performance.