Fix confusion with GraphEdge::SetShape() by renaming it
This function had the same name as the base GraphElement class method,
but did _not_ override it, as it took an argument of different type.

It seems that the intention was to only allow setting wxLineShape as
shape for the graph edges, so make it more explicit by renaming this
function to SetEdgeShape().