Reset wxShapeCanvas overlay when the canvas is repainted
Don't rely on manual calls to wxShapeCanvasOverlay::Reset(), this was
error-prone and these calls were forgotten in a number of places,
resulting in the old overlay contents being reused for subsequent drags
even after the canvas contents was modified, resulting in bad redraw
problems for any drag operations after the first one.

Instead, always reset the overlay when the canvas is repainted on the
assumption that this will be always done after modifying it.

Ideal would be to reset the overlay only when the data is actually
modified instead of doing it here because this is not quite optimal as
the overlay will also be reset even if the canvas is redrawn without
being changed, but there doesn't seem to be any single central place
from where to do this in OGL, while the extra resets don't seem to be
noticeable in practice, so keep things simple for now.