Use more appropriate OGL canvas background colour by default
Set the colour explicitly to the one used by wxListBox for its items
(e.g. white in the default non-dark theme, typically) as this is more
appropriate than the default wxPanel background colour, which is grey
under MSW.

Note that the background drawing can be, and is, in our example,
customized by handling wxEVT_ERASE_BACKGROUND in a custom handler.