Simplify and correct erasing canvas background
Use wxDC::GradientFillLinear() (which must not have been available when
this code was originally written) instead of emulating it in the code,
as this is simpler and also gives correct results even at low zoom
level, whereas the existing code left over parts of the background
unpainted in this case (probably due to rounding errors).

Also use wxDC::Clear() when not using gradient background, as this is
even simpler and must be quite a bit more efficient too.