Stop using wxClientDC
This class can't be used portably any longer, as drawing using it is not
supported by wxOSX and wxGTK when using Wayland, so stop using it and
use wxInfoDC, available in wx 3.3+ instead, to ensure that we only use
these device context objects for obtaining information about the context
(which still works in all environments) but not for updating it.

Note that compatibility with older wxWidgets versions is preserved, as
we fall back to using wxClientDC if wxInfoDC is not available.