Request full repaint on GraphCtrl resize
This fixes grid lines redrawing under Windows with wxMSW 3.2.

It should be possible to avoid gaps in the grid lines with partial
repaint too, but it doesn't seem to be worth doing this, when we already
repaint the window fully in other places and on most platforms, and even
under Windows with wxMSW 3.3, this is always the case anyhow.

This should have probably been part of 5d11195 (Always redraw the entire
visible part of the canvas, 2024-07-07).